// web/ledger-pro-ui/src/types/general-ledger-account.types.ts
export const GENERAL_LEDGER_ACCOUNT_TYPES = ['Asset', 'Liability', 'Equity', 'Revenue', 'Expense'] as const;

export type GeneralLedgerAccountType = typeof GENERAL_LEDGER_ACCOUNT_TYPES[number];

export interface GeneralLedgerAccount {
    id: number;
    name: string;
    description?: string;
    accountType: GeneralLedgerAccountType;
}